requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

python-telegram-bot == 20.5

eyed3 == 0.9.7

spotipy == 2.24.0

requests == 2.32.2

yt-dlp == 2024.10.7
